### 2022-04-15 #4319 @editorialbot
[PRE REVIEW]: MyPTV: A Python Package for 3D Particle Tracking
**Submitting author:** @ronshnapp (<a href="http://orcid.org/0000-0001-7495-8420">Ron Shnapp</a>)
**Repository:** https://github.com/ronshnapp/MyPTV
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v0.2
