### 2022-04-15 #4319 @Kevin-Mattheus-Moerman
@ronshnapp thanks for this submission. I am currently checking if this work is in scope for JOSS. In the mean time can you:

- [ ]  work on those potentially missing/invalid DOI's? You can call `@editorialbot  check references` to check the references again and use `@editorialbot generate pdf` to update the paper.
- [ ] add an [OSI approved license](https://opensource.org/licenses). The best thing to do is to add a plane LICENSE file to your repository, that way GItHub and our systems pick up on it easily too.
