### 2022-04-20 #4319 @arfon
@ronshnapp - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area (@kyleniemeyer) is already rather busy.
