### 2022-04-21 cc3af89
In the segmentation workflow, for a single image now we specify the image name in the file
