+++ example/workflow.py
@@ -46,7 +46,7 @@ class workflow(object):
         
         # perform the wanted action:
         if action is None:
-            print('Started workflow wi no particular action.')
+            print('Started workflow with no particular action.')
             
         elif action != None:
             
@@ -119,9 +119,16 @@ class workflow(object):
         '''
         Fetches a parameter value from the self.params DataFrame.
         '''
+        if act not in set(self.params['operation']):
+            raise ValueError('Cant find action %s in the parameter file.'%act)
+        
         par_seg = self.params[self.params['operation']==act]
+        
+        if param not in set(par_seg['param']):
+            msg = 'Cant find the %s -> %s in the parameters file.'%(act,param)
+            raise ValueError(msg)
+        
         return par_seg[par_seg['param']==param]['value'].iloc[0]
-    
     
     
     
@@ -293,6 +300,7 @@ class workflow(object):
         plot_res = self.get_param('segmentation', 'plot_result')
         save_name = self.get_param('segmentation', 'save_name')
         ROI = self.get_param('segmentation', 'ROI')
+        single_img_name = self.get_param('segmentation', 'single_image_name')
         
         
         # reading preprepared mask
@@ -303,7 +311,10 @@ class workflow(object):
         allfiles = os.listdir(dirname)
         n_ext = len(ext)
         image_files = sorted(list(filter(lambda s: s[-n_ext:]==ext, allfiles)))
-        image0 = imread(os.path.join(dirname,image_files[0]))
+        if single_img_name in image_files:
+            image0 = imread(os.path.join(dirname,single_img_name))
+        else:
+            image0 = imread(os.path.join(dirname,image_files[0]))
         
         # preparing a mask using the given ROI
         if ROI is not None:
@@ -311,7 +322,6 @@ class workflow(object):
             mask_ROI = zeros(image0.shape)
             mask_ROI[ROI[2]:ROI[3]+1, ROI[0]:ROI[1]+1] = 1
             mask = mask * mask_ROI
-        
         
         # segmenting the image if there are more than 1 frames
         if N_img is None or N_img>1:
@@ -340,10 +350,15 @@ class workflow(object):
             print('Done.')
         
         
-        # segmenting the image if is only 1 frames
+        # segmenting the image if there is only 1 frames
         if N_img == 1:
-            print('starting segmentation on a single image.')
-            print(os.path.join(dirname,image_files[0]))
+            print('\n','starting segmentation on a single image.')
+            if single_img_name not in image_files:
+                in_ = os.path.join(dirname,single_img_name)
+                msg = 'Image %s not found in the directory.'%in_
+                raise ValueError(msg)
+            
+            print('\n','segmenting image: %s'%single_img_name)
             particleSegment = particle_segmentation(image0, 
                                                     sigma=sigma, 
                                                     threshold=threshold, 
@@ -560,7 +575,7 @@ class workflow(object):
         print('\n', 'Done.')
         
         
-        
+#%%
         
         
         
