+++ example/params_file.yml
@@ -9,10 +9,11 @@
 - segmentation:
     Number_of_images: 1
     images_folder: 'Calibration'
+    single_image_name: cal2.tif
     image_extension: '.tif'
     mask: 1.0
     ROI: 227, 1050, 156, 932
-    plot_result: False
+    plot_result: True
     threshold: 14.0
     blur_sigma: 0.4 
     local_filter: 20
