### 2022-04-21 4e7cc3c
added the new single image filename workflow to the manual
