### 2022-04-22 d777085
small changes to the matching function
