+++ myptv/cal_point_gui.py
@@ -1,0 +1,355 @@
+# -*- coding: utf-8 -*-
+"""
+Created on April 23, 2022
+
+@author: Ron
+
+A gui for manually marking points on a calibration target. 
+"""
+
+
+from PIL import Image, ImageTk
+from tkinter import Label, Canvas, LabelFrame, Entry, Tk, Scrollbar, Button
+
+
+
+
+
+class cal_point_gui(object):
+    '''
+    This is a Tkinter based graphical user interface that can be used to mark
+    points on a static image, give their lab space coordinates, and then save
+    the data as a text file in the format used by the calibration processes.
+    '''
+    
+    
+    def __init__(self, image_name, savename):
+        '''
+        input: 
+        
+        image_name - the calibration image used choose the points
+        
+        savename - the name of the text file in which the extracted data will
+                   be saved.
+        '''
+        self.fname = savename
+        self.image_name = image_name
+        self.xy_marked = (-1, -1)
+        self.point_list = []      # <-- list of points to save
+        self.point_markers = []   # <-- position of crosses
+        self.z = 1.0              # <-- Zoom level
+        
+        # set the window
+        self.root = Tk()
+        self.root.geometry('1000x700+320+70')
+        self.root.title('MyPTV: calibration point marking GUI')
+        
+        # tracking_dic = {}
+        # point_markers = []
+        
+        # load the image
+        image = Image.open(self.image_name)
+        image_size = image.size
+        photo = ImageTk.PhotoImage(image)
+        
+        
+        # buttons frame
+        button_label = LabelFrame(self.root, padx=10, pady=10, width=100)
+        button_label.grid(row=0, column=1, padx=(10), pady=10, sticky='nsew')
+        
+        # place the image inside a canvas in a frame
+        self.board = Label(image=photo)
+        frame = LabelFrame(self.root, bg='grey', width=700, height=450,
+                           padx=(5), pady=5)
+        frame.grid(row=0, column=0, padx=(5), pady=5, sticky='nsew')
+        #frame.place(x=120, y=30)
+        
+        sr = (0,0,image_size[0],image_size[1])
+        self.canvas = Canvas(frame, scrollregion=sr, height=450, width=700)       
+        self.canvas.grid(row=0,column=1, sticky='ewns', padx=(5), pady=5)
+        #self.canvas.focus()
+        
+        # add scrolbars to the image
+        self.hbar=Scrollbar(frame,orient='horizontal')
+        self.hbar.grid(row=1,column=1, sticky='ew')
+        self.hbar.config(command=self.canvas.xview)
+        self.vbar=Scrollbar(frame,orient='vertical')
+        self.vbar.grid(row=0,column=0, sticky='nse')
+        self.vbar.config(command=self.canvas.yview)
+        
+        self.canvas.config(xscrollcommand=self.hbar.set,
+                            yscrollcommand=self.vbar.set)
+        self.canvas.create_image(0, 0, image=photo, anchor='nw')
+        
+        # board.place(x=140, y=30)
+        
+        # setup click bottun
+        self.canvas.bind("<Button-1>", self.location_handler)
+        
+        
+        # selection indicators
+        select_frame = LabelFrame(button_label, padx=2, pady=2)
+        select_frame.grid(row=1, column=0, columnspan=2, sticky='ew',
+                          padx=2, pady=2)
+        self.xloc = Label(select_frame, text='x image:', padx=2, pady=2)
+        self.yloc = Label(select_frame, text='y image:', padx=2, pady=2)
[…]
