### 2022-04-24 2ca4005
added the calibration point marking gui to the workflow
