### 2022-04-24 026c12a
updated the user manual
