+++ user_manual/user_manual.tex
@@ -271,15 +271,133 @@ The \texttt{params\_file.yml} outlines all the parameters needed in order to run
 
 
 
-\subsection{Calibration}
+\subsection{Calibration guide}
 
-A good calibration is key to having success in your PTV experiment! Thus, follow this guide to calibrate your camera using MyPTV.
-
-
-First, we need to prepare a \textit{target file}. This is a text file which lists the lab-space coordinates of all the points on our calibration target according to the format shown in Fig.~\ref{fig:target_file}. Prepare such a file and save it in the experiment's calibration folder.
+A good calibration is key to having success in your PTV experiment! Thus, follow this guide to calibrate your camera using MyPTV. 
 
 
-Then, each camera in MyPTV is calibrated separately. Thus, to calibrate a camera, we make sure that we have a calibration image ready (as seen in Fig.~\ref{fig:cal_img}, for example). Then, we perform an initial "rough" calibration, and finish of with the final calibration. 
+Camera calibration consists of several steps in which we determine the external ($\vec{O}$, $\vec{\theta}$) and internal ($f$, $x_h,\,y_h$, $[E]$) parameters of our camera system. Each camera in MyPTV is calibrated separately using a \textit{calibration image} - an image of an object with marked points on it whose coordinates are known to us, a \textit{calibration target}. Therefore, to calibrate a camera, we make sure that we have a calibration image ready, as seen, for example, in Fig.~\ref{fig:cal_img}. 
+
+Importantly, since the goal is finding the cameras' positions, once a calibration image was taken in an experiment it is critical that the cameras are not moved. Also, the calibration target needs to be in the same medium in which the experiment will be performed to ensure that any refraction is corrected by the error terms (e.g. if you are using a tank of water, place the calibration target inside the tank of water). 
+
+
+A schematic block diageam for the calibration process of each camera is given in Fig.~\ref{fig:calibsteps}. The calibration is divided into two sub-processes: 1) initial calibration, in which we use only 6 calibration points to obtain an approximate calibration estimation, and, 2) final calibration, in which we use all the calibration points on the calibration target to improve the accuracy of the initial calibration. The various steps of the process are performed in MyPTV through the workflow script, and detailed instructions are given below. 
+
+
+\begin{figure}
+	\centering
+	\includegraphics[width=10cm]{Calibration_steps.pdf}
+	\caption{A block diagram representing the steps to follow in the calibration of each camera using MyPTV. \label{fig:calibsteps}}
+\end{figure}
+
+
+
+
+
+\subsubsection{Preparing the parameters in the \texttt{params\_file}}
+
+As in any step of MyPTV, we first make sure that the \texttt{params\_file.yaml} in our experiment folder has the correct parameters, namely file names under the calibration tab; see Tab.~\ref{tab:cal_paramsfile}. Each camera is calibrated separately, and so, for the calibration of each camera we need to change the file names in \texttt{params\_file.yaml} appropriately (alternatively, one could have separate parameter files for each camera).  
+
+\begin{table}
+	\centering
+	\caption{The \texttt{params\_file.yml} parameters for the calibration step. All paths to files are relative to the \texttt{workflow.py} script. \label{tab:cal_paramsfile}}
+	\begin{tabular}{l m{10cm}}
+		\hline
+		Parameter & Description\\
+		\hline
+		\texttt{camera\_name} & the name of the camera to be calibrated \\
+		
+		\texttt{calibration\_points\_file} & path to the file that lists the correlated lab space and image space coordinates \\
+		
+		\texttt{target\_file} & path to the file that lists the calibration target's lab points \\
+		
[…]
