+++ setup.py
@@ -9,9 +9,9 @@ from setuptools import find_packages, setup
 setup(
     name='myptv',
     packages=find_packages(include=['myptv']),
-    version='0.2.1',
+    version='0.2.2',
     description='A 3D Particle Tracking Velocimetry library',
-    install_requires=['numpy', 'scipy', 'scikit-image','pandas','matplotlib','pyyaml'],
+    install_requires=['numpy', 'scipy', 'scikit-image','pandas','matplotlib','pyyaml', 'tk', 'Pillow'],
     author='Ron Shnapp',
     author_email='ronshnapp@gmail.com',
     license='MIT',
