+++ environment.yml
@@ -16,3 +16,5 @@ dependencies:
   - scikit-image
   - scipy
   - pyyaml
+  - pillow
+  - tk
