+++ user_manual/bib_myPTV.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 @Article{Virant1997,
   author     = {M. Virant and T. Dracos},
   journal    = {Measurement},
@@ -13,20 +14,19 @@
   url        = {http://iopscience.iop.org/article/10.1088/0957-0233/8/12/017/meta},
 }
 
-@Article{Mass1993,
-  Title                    = {Particle Tracking Velocimetry in Three-Dimensional Flows Part I: Photogrammetric Determination of Particle Coordinates},
-  Author                   = {Mass, H. G. and Gruen, D. and Papantoniou, D.},
-  Journal                  = {Experiments in Fluid},
-  Year                     = {1993},
-  Pages                    = {133-146},
-  Volume                   = {15},
-
-  Bdsk-url-1               = {https://link.springer.com/article/10.1007/BF00190953#citeas},
-  Bdsk-url-2               = {https://doi.org/10.1007/BF00190953},
-  Doi                      = {https://doi.org/10.1007/BF00190953},
-  Owner                    = {ron},
-  Timestamp                = {2018.02.28},
-  Url                      = {https://link.springer.com/article/10.1007/BF00190953#citeas}
+@Article{Maas1993,
+  author     = {Maas, H. G. and Gruen, D. and Papantoniou, D.},
+  journal    = {Experiments in Fluid},
+  title      = {Particle Tracking Velocimetry in Three-Dimensional Flows Part I: Photogrammetric Determination of Particle Coordinates},
+  year       = {1993},
+  pages      = {133-146},
+  volume     = {15},
+  bdsk-url-1 = {https://link.springer.com/article/10.1007/BF00190953#citeas},
+  bdsk-url-2 = {https://doi.org/10.1007/BF00190953},
+  doi        = {https://doi.org/10.1007/BF00190953},
+  owner      = {ron},
+  timestamp  = {2018.02.28},
+  url        = {https://link.springer.com/article/10.1007/BF00190953#citeas},
 }
 
 @Article{Bourgoin2020,
