+++ user_manual/user_manual.tex
@@ -2,7 +2,7 @@
 
 
 \usepackage[T1]{fontenc}
-\usepackage[utf8]{inputenc}
+% \usepackage[utf8]{inputenc}
 \usepackage[english]{babel}
 \usepackage[english]{isodate}
 \usepackage[parfill]{parskip}
@@ -11,8 +11,9 @@
 \graphicspath{{./figs/}}
 \usepackage{amsmath}
 \usepackage{amssymb}
-\usepackage{cite}
+%\usepackage{cite}
 \usepackage{array}
+\usepackage{natbib}
 
 
 \usepackage{xcolor}
@@ -91,10 +92,10 @@
 \subsection{3D-PTV principles}
 
 
-The 3D-PTV method is used to measure trajectories of particles in 3D space. It utilities the principles of stereoscopic vision in order to reconstruct 3D positions of particles from images taken from several angles. A scheme of a typical 3D-PTV experiment using a four camera system is shown in Fig.~\ref{fig:3dptv_exp}. The "work horse" behind the 3D-PTV method is the co-linearity condition, the 3D model. In principle, if we know what is the position and what is the orientation of the camera in 3D space ($O'$ and $\theta$ in Fig.~\ref{fig:3dmodel}), we can use the pin-hole camera model to relate the image space coordinates of a particle ($\eta,\, \zeta$ in Fig.~\ref{fig:3dmodel}) to the ray of light connecting the imaging center and the particle. Then, if we have more than one camera, the particle will be located at the intersection of the two rays. Detailed information is given in \cite{Virant1997, Mass1993}.  
+The 3D-PTV method is used to measure the trajectories of particles in 3D space. It utilizes the principles of stereoscopic vision to reconstruct 3D positions of particles from images taken from several angles. A scheme of a typical 3D-PTV experiment using a four camera system is shown in Fig.~\ref{fig:3dptv_exp}. The "work horse" behind the 3D-PTV method is the co-linearity condition, the 3D model. In principle, if we know what is the position and what is the orientation of the camera in 3D space ($O'$ and $\theta$ in Fig.~\ref{fig:3dmodel}), we can use the pin-hole camera model to relate the image space coordinates of a particle ($\eta,\, \zeta$ in Fig.~\ref{fig:3dmodel}) to the ray of light connecting the imaging center and the particle. Then, if we have more than one camera, the particle will be located at the intersection of the two rays. Detailed information is given in \cite{Virant1997, Maas1993}.  
 
 
-\begin{figure}[h!]
+\begin{figure}[!ht]
 	\centering
 	\subfloat[]{\label{fig:3dptv_exp}
 		\includegraphics[width=6.5cm]{3D_PTV_acquisition.pdf}}
@@ -166,7 +167,7 @@ where $[E]$ is a $3\times5$ matrix that holds the correction coefficients; the l
 
 
 
-\begin{table}[h]
+\begin{table}[!ht]
 	\centering
 	\caption{Description of mathematical notation. \label{tab1:mathdesc}}
 	\begin{tabular}{p{5em} p{30em}}
@@ -281,7 +282,7 @@ Camera calibration consists of several steps in which we determine the external 
 Importantly, since the goal is finding the cameras' positions, once a calibration image was taken in an experiment it is critical that the cameras are not moved. Also, the calibration target needs to be in the same medium in which the experiment will be performed to ensure that any refraction is corrected by the error terms (e.g. if you are using a tank of water, place the calibration target inside the tank of water). 
 
 
[…]
