+++ myptv/tracking_mod.py
@@ -616,8 +616,17 @@ class tracker_nearest_neighbour(object):
         
         
         
+        
+        
+# =============================================================================
+#                        score optimizing tracker
+# =============================================================================
+        
+        
+        
+        
 
-class cost_minimizing_tracker():
+class score_optimizing_tracker():
     '''
     This class implements a tracker that uses cost function minimization.
     First, for each particle we obtain candidate links in the next frame. The
@@ -672,7 +681,7 @@ class cost_minimizing_tracker():
             self.particles[tm] = []
             p_ = data[data[:,-1]==tm]
             for i in range(p_.shape[0]):
-                p = array(list(p_[i,[0,1,2,-1]]))
+                p = array([-1]+ list(p_[i,[0,1,2,-1]]))
                 self.particles[tm].append(p)
         
         # a dictionary of KDtrees based on the frame, so we dont have to make
@@ -694,6 +703,23 @@ class cost_minimizing_tracker():
         
         
         
+    def track_frames(self):
+        '''
+        This function performs all the tracking steps, and sets up the 
+        connected particles results as self.results
+        '''
+        self.get_all_candidate_links()
+        self.make_candidate_chains()
+        self.sort_chains(score=None)
+        self.prune_trajectories()
+        
+        self.results = []
+        for i in range(len(self.trajs)):
+            self.trajs[i][:,0] = i
+            for j in range(len(self.trajs[i])):
+                self.results.append(self.trajs[i][j,:])
+                
+        
     def get_all_candidate_links(self):
         '''
         A function that fills self.candidate_links withh all possible 
@@ -702,7 +728,11 @@ class cost_minimizing_tracker():
         for k in self.particles.keys():
             for i in range(len(self.particles[k])):
                 self.get_neighbouring_candidates(self.particles[k][i], i)
-            
+                
+        for k in self.particles.keys():
+            for i in range(len(self.particles[k])):
+                self.get_velocity_projected_candidates(self.particles[k][i], i)
+        
         
         
         
@@ -727,19 +757,69 @@ class cost_minimizing_tracker():
             tree = self.trees[frm+1]
         
         else:
-            particles_ip1 = array(self.particles[frm+1])[:,:3]
+            particles_ip1 = array(self.particles[frm+1])[:,1:4]
             tree = KDTree(particles_ip1)
             self.trees[frm+1] = tree
-            
-        # add the candidate links to the dictionary
+        
+        # make sure there is an entry for our frame in self.candidate_links 
         if frm not in self.candidate_links.keys():
-            self.candidate_links[frm] = []
-        candidate_indexes = tree.query_ball_point(particle[:3], self.d_max)
+            self.candidate_links[frm] = set([])
+        
+        # add the candidate links to the dictionary
+        candidate_indexes = tree.query_ball_point(particle[1:4], self.d_max)
         for ci in candidate_indexes:
-            self.candidate_links[frm].append( (p_index, ci, frm) )
+            self.candidate_links[frm].add( (p_index, ci, frm) )
             
         return
         
+    
+    
+    def get_velocity_projected_candidates(self, particle, p_index):
+        '''
+        This function searcher linking candidates of given trajectories
+        in regions that are close to the projection of the paritcle into
+        the next frame, assuming the particle has a constant velocity.
+        Specifically, if x(i) is the position at frame i, that we are looking
+        for neighbours of the point 
+        x(i+1) = x(i) + x(i)-x(i-1) = 2x(i) - x(i-1)
+        
[…]
