+++ myptv/match_track_iter_mod.py
@@ -1,0 +1,316 @@
+# -*- coding: utf-8 -*-
+"""
+Created on April 27th 2022
+
+@author: ron
+
+Contains a class to iteratively perform matching and tracking and slowly 
+depleat blobs and track more particles
+
+"""
+
+from numpy import loadtxt, array, savetxt
+from pandas import read_csv
+#from myptv.particle_matching_mod import match_blob_files
+#from myptv.tracking_mod import score_optimizing_tracker
+from particle_matching_mod import match_blob_files
+from tracking_mod import score_optimizing_tracker
+
+
+
+
+class iterative_matcher_tracker(object):
+    
+    
+    
+    def __init__(self, blob_fnames, img_system, ROI, voxel_size, 
+                 max_blob_dist, mean_flow, d_max, dv_max, max_err, 
+                 threshold_traj_len, frame_start=None, frame_end=None):
+        
+        # load the blobs
+        self.blobs = []
+        for fn in blob_fnames:
+            #self.blobs.append(loadtxt(fn))
+            self.blobs.append(array(read_csv(fn, sep='\t', header=None)))
+        
+        # discard blobs out of the given frame range
+        if frame_start is not None:
+            for i in range(len(self.blobs)):
+                lst = self.blobs[i]
+                self.blobs[i] = lst[lst[:,-1]>=frame_start]
+        
+        if frame_end is not None:
+            for i in range(len(self.blobs)):
+                lst = self.blobs[i]
+                self.blobs[i] = lst[lst[:,-1]<=frame_end]
+        
+        
+        # ====================================================
+        #for i in range(len(self.blobs)):
+        #    lst = self.blobs[i]
+        #    self.blobs[i] = lst[(lst[:,0]>500)&(lst[:,1]>400)]
+        # ====================================================
+        
+        
+        self.orig_blobs = self.blobs.copy()
+        
+        # unpack matching and tracking parameters 
+        self.imsys = img_system
+        self.ROI = ROI
+        self.voxel_size = voxel_size
+        self.max_blob_dist = max_blob_dist
+        self.max_err = max_err
+        self.mean_flow = mean_flow
+        self.d_max = d_max
+        self.dv_max = dv_max
+        self.threshold_traj_len = threshold_traj_len
+        
+        
+        # a list that will hold the trajectories
+        self.trajectories = []
+        
+        self.count_blobs = 0
+        
+        
+        
+    def match_and_track(self, voxel_size = None):
+        '''
+        performas a round of matching and then tracking the unused blobs.
+        '''
+        
+        print('\n', 'number of available blobs:',
+              sum([len(lst) for lst in self.blobs]), '\n')
+        
+        if voxel_size is not None:
+            self.voxel_size = voxel_size
+        
+        # Doing the matching 
+        mbf = match_blob_files([], self.imsys, self.ROI, self.voxel_size, 
+                               self.max_blob_dist, blobs = self.blobs, 
+                               max_err = self.max_err)
+        mbf.get_particles()
+        particles = mbf.return_particles()
+        
+        
+        # Tracking the newly obtained particles
+        self.sot = score_optimizing_tracker([], mean_flow = self.mean_flow, 
+                                       d_max=self.d_max, dv_max=self.dv_max,
+                                       particles = particles)
+        self.sot.track_frames()
+        
+        
+        # a nested list of dictionaries that hold the indexes of the blobs 
+        # that were used to make trajectories. The sublist index is the camera 
[…]
