### 2022-05-01 ea74e43
not the tracking algorithm saves also data for blobs
