### 2022-05-01 647b051
explained the new trajectory file format in the manual
