+++ example/workflow.py
@@ -64,6 +64,9 @@ class workflow(object):
             
             elif action == 'calibration_point_gui':
                 self.calibration_point_gui()
+                
+            elif action == 'calibration_with_particles':
+                self.calibration_with_particles()
             
             elif action == 'match_target_file':
                 self.match_target_file()
@@ -302,6 +305,83 @@ class workflow(object):
         print('\n', 'Starting calibration point segmentation GUI', '\n')
         gui = cal_point_gui(cal_image, blob_file)
     
+    
+    
+    def calibration_with_particles(self):
+        '''
+        This starts the calibrate with particles sequence
+        '''
+        from myptv.imaging_mod import camera
+        from myptv.calibrate_mod import calibrate_with_particles
+        from  matplotlib.pyplot import subplots
+        
+        # fetch parameters from the file
+        camera_name =  self.get_param('calibration_with_particles',
+                                      'camera_name')
+        resolution = self.get_param('calibration_with_particles',
+                                      'resolution')
+        traj_filename = self.get_param('calibration_with_particles',
+                                      'traj_filename')
+        cam_number = self.get_param('calibration_with_particles',
+                                      'cam_number') 
+        blobs_fname = self.get_param('calibration_with_particles',
+                                      'blobs_fname')
+        min_traj_len = self.get_param('calibration_with_particles',
+                                      'min_traj_len')
+        max_point_number = self.get_param('calibration_with_particles',
+                                      'max_point_number')
+        print('\n', 'starting calibration with particles')
+        
+        # setting up a camera instance            
+        cam = camera(camera_name, resolution)
+        cam.load('./')
+        
+        
+        # set up the calibration object
+        cal_with_p = calibrate_with_particles(traj_filename, cam, cam_number, 
+                                      blobs_fname, min_traj_len=min_traj_len,
+                                      max_point_number=max_point_number)
+        
+        
+        cal = cal_with_p.get_calibrate_instance()
+        print('\n', 'ready to calibrate')
+        print('initial error: %.3f pixels'%(cal.mean_squared_err()))
+        print('')
+        
+        user = True
+        print('Starting calibration sequence:')
+        while user != '9':
+            print("enter '1' for external parameters calibration")
+            print("enter '2' for internal correction ('fine') calibration")
+            print("enter '3' to show current camera external parameters")
+            print("enter '4' to plot the calibration points' projection")
+            print("enter '8' to save the results")
+            print("enter '9' to quit")
+            user = input('')
+            
+            if user == '1':
+                print('\n', 'Iterating to minimize external parameters')
+                cal.searchCalibration(maxiter=2000)
+                err = cal.mean_squared_err()
+                print('\n','calibration error: %.3f pixels'%(err),'\n')
+            
+            if user == '2':
+                print('\n', 'Iterating to minimize correction terms')
+                cal.fineCalibration()
+                err = cal.mean_squared_err()
+                print('\n','calibration error:', err,'\n')
+                
+            if user == '3':
+                print('\n', cam, '\n')
+            
+            if user == '4':
+                fig, ax = subplots()
+                cal.plot_proj(ax=ax)
+                
+            if user == '8':
+                print('\n', 'Saving results')
+                cam.save('.')
+        
     
     
     def do_segmentation(self):
