+++ user_manual/user_manual.tex
@@ -57,7 +57,7 @@
 
 \begin{minipage}{10cm}
 	{\small 
-	Vesrion: 0.2.3 \\
+	Vesrion: 0.3.0 \\
 	Last updated: \today \\
 	Github repository: \url{https://github.com/ronshnapp/MyPTV} \\
 	Contac: ronshnapp@gmail.com}
