### 2022-05-02 fcb097e
added the calibration with particles to the user manual
