+++ myptv/extended_temporal_linking_mod.py
@@ -1,0 +1,270 @@
+# -*- coding: utf-8 -*-
+"""
+Created on May 2022
+
+@author: ron
+
+This module is used to extend the matching and tracking of particles using
+temporal data, using the ideas developed by Schants et al. 2016 in their 
+Shake the Box method.
+
+"""
+
+
+
+
+from scipy.spatial import KDTree
+from scipy.optimize import minimize
+from itertools import combinations, product
+from numpy import append as npappend
+
+
+class extended_linking(object):
+    '''
+    This class takes in a trajectory that its last point is on frame i, and
+    by looking back at the blob files for frame i+1, tries to find whether it
+    may have a connection that was missed in the matching stage.
+    We do this by 1) estimating the particle's position at frame i+1 using its 
+    velocity, 2) projecting this position to camera space coordinates at all
+    cameras, 3) taking the blobs that are nearest to the projection in each
+    camera, and 4) attemptin to stereo match these nearest neighbouring blobs. 
+    If this is successfull, the new particle sample is added to the list of 
+    trajectories with the same trajectory number ID as the one given initially.
+    '''
+    
+    def __init__(self, imsys, particles_dic, pixel_neighbourhood_size, dv_max,
+                 max_err=1e9,
+                 reverse_eta_zeta=False):
+        '''
+        imsys - an instance of the img_system class
+        
+        particles_dic - a dictionary with keys being camera names, and values
+                        being lists of blobs that belong to the frame i+1 and
+                        camera.
+                        
+        max_err - float, the maximum allowable triangulation error in the 
+                  stereo matching
+                  
+        reverse_eta_zeta - should the order of image space coordinates be
+                           reversed.
+        '''
+        
+        self.imsys = imsys
+        self.pd = particles_dic
+        self.reverse_eta_zeta = reverse_eta_zeta
+        self.max_err = max_err
+        self.r = pixel_neighbourhood_size
+        self.dv_max = dv_max
+        
+        self.trees = dict([(k, KDTree(self.pd[k][:,:2])) 
+                           for k in self.pd.keys()])
+    
+    
+    
+    def nearest_blobs_distance(self, x):
+        '''
+        Given a location x, this returns the sum of the distances between the
+        projections of x on all cameras and their nearest neighbours
+        '''
+        dist_list = []
+        for cam in self.imsys.cameras:
+            if self.reverse_eta_zeta==False:
+                proj = cam.projection(x)[::-1]
+            else:
+                proj = cam.projection(x)
+            dist_list.append( self.trees[cam.name].query([proj])[0][0] )
+        
+        return dist_list
+
+        
+    
+    def optimize_particle_position(self, x0, bounds):
+        '''
+        Given an initial position x0, this function tries to "shake" it
+        such that its nearest blobs distance in image space is minimized.
+        '''
+        cost_func = lambda a: sum(self.nearest_blobs_distance(a))
+        
+        xmin = minimize(cost_func, x0, bounds=bounds)
+        return xmin
+    
+    
+    
+    
+    def search_for_stereo_matching(self, tr):
+        '''
+        Given a point in lab space, x, this will search for the blobs that 
+        are this points' neighbourhood in image space and will stereo 
+        match them.
+        '''
+        
+        # calculate the projected position
+        x = tr[-1,1:4] + tr[-1,1:4] - tr[-2,1:4]   
+        
+        # for each camera, get the neighbouring blobs and put them in lists,
+        # according to their camera
+        candidates_by_cams = []
[…]
