+++ Readme.md
@@ -1,5 +1,5 @@
-April 13, 2022
-Version: 0.3.0
+May 12, 2022
+Version: 0.3.1
 
 # MyPTV
 
