### 2022-05-14 0c361c0
Now one can choose from one of two segmentation methods - using dilation or the originally used blob labeling
