+++ myptv/segmentation_mod.py
@@ -8,11 +8,13 @@ Created on Fri Dec  7 18:02:07 2018
 contains a class for segmentation of circular particles
 """
 
-from numpy import zeros, savetxt
-from scipy.ndimage import gaussian_filter
+from numpy import zeros, ones, savetxt, meshgrid
+from numpy import sum as npsum
+
 from skimage.io import imread
-from numpy import ones
+
 from scipy.signal import convolve2d
+from scipy.ndimage import gaussian_filter
 from scipy.ndimage.measurements import label, find_objects
 
 
@@ -20,13 +22,51 @@ class particle_segmentation(object):
     '''a class for segmenting out particles (blobs) for a given image'''
     
     
-    def __init__(self, image, sigma=1.0, threshold=10, mask=1.0,
+    def __init__(self, image, particle_size, sigma=1.0, threshold=10, mask=1.0,
                  local_filter = 15,
                  min_xsize=None, max_xsize=None,
                  min_ysize=None, max_ysize=None,
-                 min_area=None, max_area=None):
+                 min_area=None, max_area=None,
+                 method='dilation'):
+        '''
+        inputs - 
+        
+        image - the image (matrix) to be analyzed
+        
+        particle_size - the expected size of particles; this input is only
+                        used if method='dilation'.
+        
+        sigma - float, standard deviation of the gaussian blur filter. If this 
+                is None, then the bluring filter is not applied.
+        
+        threshold - int, the grey value above which pixels are considered a 
+                    particle. Note that the threshold is performed after the
+                    blur and local mean subtraction filters are applied.
+        
+        mask - either array of 0 and 1 with the same shape of imae, or 1. If 
+               this is an array of 0 and 1, this is taken to choose regions 
+               of interest in the analyzed image (regions with 0 are 
+               discarded).
+        
+        local_filter - int, the window size of the local mean subtraction 
+                       filter. In this is None then the filter is not applied.
+                       
+        min/max_( ) - size and area filters for the discovered blobs. If None
+                      then filters are not applied.
+                      
+        method - string. The method used for labeling the blobds. Can be 
+                 either 'dilation' or 'labeling'. In dilation, local maxima 
+                 are sought in the image and the blobs are considered to be
+                 regions of size 'particle_size' around these maxima. In the 
+                 'labeling' method, the threshold is performed and blobs are
+                 considered to be connected bright pixels. The dilation method
+                 is better at distinguishing close particles (saddle points in 
+                 the brihtness) if their size is given correctly. The 
+                 'labeling' method may be better for non spherical particles.
+        '''
         
         self.im = image
+        self.p_size = particle_size
         self.sigma = sigma
         self.th = threshold
         self.mask = mask
@@ -34,7 +74,13 @@ class particle_segmentation(object):
         self.area_limits = (min_area, max_area)
         self.loc_filter = local_filter
         
-    
+        if method not in ['dilation', 'labeling']:
+            raise ValueError('method "%s" unknown.'%method)
+        
+        self.method = method
+        
+        
+
     def local_filter(self, image):
         '''returns a new image where the local mean neighbourhood of
         each pixel is subtracted.'''
@@ -46,12 +92,15 @@ class particle_segmentation(object):
         new_im = new_im.astype('uint8')
         return new_im
         
-        
-    def get_binary_image(self):
-        '''Will mark pixels in the image as background and foreground 
-        (particles). We blur the image with a Gaussian
-        filter, and look for regions brighter than a global threshold 
-        level.'''
+    
+    
+    def process_image(self):
+        '''
[…]
