### 2022-05-14 87ab87c
adding segmentation new methods to workflow
