### 2022-05-15 754f1a3
fixed duplicates issue in the dilation method; Now segmentation plotting shows the processed image
