### 2022-05-15 bbc195e
updated the new segmentation methods in the manual; transitioned to V0.4.0
