### 2022-05-17 89092e3
added a median filter to segmentation
