+++ user_manual/user_manual.tex
@@ -535,8 +535,13 @@ In this step we extract the particles' image-space coordinates from the images a
 \begin{center}
 	
 \fcolorbox{white}{bcolor}{
-	\begin{minipage}{0.95\textwidth}
-		\textbf{How segmentation is performed:} Segmentation in MyPTV is performed in one of two methods: \textit{labeling}, or \textit{dilation}. 
+	\begin{minipage}[c][][c]{0.95\textwidth}
+		
+		\vspace{.1cm}
+		
+		\textbf{How segmentation works:} Segmentation in MyPTV is performed in one of two methods: \textit{labeling}, or \textit{dilation}, and includes several optional image processing steps. 
+		
+		\vspace{.2cm}
 		
 		\begin{enumerate}
 			\item In the labeling method, we first choose all pixels in an image whose grey value is higher than a given threshold. Then, such pixels that are touching each other are considered to be "blobs", so we group them together. For each blob, we calculate a \textit{center of mass}, being the grey value weighted average of the blob, a \textit{bounding box} size, and a \textit{mass}, being the sum of grey values of the blobs' pixels.
@@ -544,7 +549,17 @@ In this step we extract the particles' image-space coordinates from the images a
 			\item In the dilation method, we are looking for pixels that their grey value is higher than that of all their neghbours within a box of a given size (\texttt{particle\_size}), and thet their grey value is higher than a given threshold. After that, all the neighbouring pixels inside the particle size box are considered to be a blob. We then calculate the center of mass of the blob to achieve a better estimation of its position, thus redifining the blob. This process is iterated a maxinum of 3 times untill convergence is obtained. Finally, we calculate the same blob's \textit{center of mass}, \textit{bounding box} size, \textit{mass}.
 		\end{enumerate}
 		
-		To imrove the results of segmenting over non-ideal images, MyPTV can be set to apply imae processing algorithms betfore the blob extraction: a gaussian blur filter, and local mean subtraction filter, and a mask. The first two filters mmay increase the computational time, however they may be necessary depending on the experimental conditions.
+		\vspace{.2cm}
+		
+		To imrove the results of segmenting over non-ideal images, MyPTV can be set to apply image processing before the blob extraction. The fiters that can be used are a gaussian blur filter, a median noise removal filter, local mean subtraction filter, and a mask. The first three filters may increase the computational time, however they may be necessary depending on the experimental conditions, such as uneven illumination or noisy images. Examples of the results of using several combinations of the filters on a calibration image is shown below. 
+		  
+		 \vspace{.2cm}
+		  
+	\begin{center}
+		\includegraphics[width=0.9\textwidth]{filters_effect.pdf}
+	\end{center}
+
+		 \vspace{.1cm}
 		  
 	\end{minipage}
 }
@@ -580,9 +595,11 @@ In this step we extract the particles' image-space coordinates from the images a
 		
 		\texttt{threshold} & the brightness value for which a point is considered a particle after a local mean subtraction is performed \\[.3em]
 		
-		\texttt{blur\_sigma} & the standard deviation of a Gaussian blur filter \\[.3em]
+		\texttt{median} & the intger window size of a median noise removal filter; set to \texttt{None} in order to not apply the filter \\[.3em]
+		
+		\texttt{blur\_sigma} & the float standard deviation of a Gaussian blur filter; set to \texttt{None} in order to not apply the filter \\[.3em]
 		 
-		\texttt{local\_filter} & the size of a local mean subtraction filter \\[.3em]
+		\texttt{local\_filter} & the integer widow size of a local mean subtraction filter; set to \texttt{None} in order to not apply the filter \\[.3em]
 		
 		\texttt{min\_xsize} & minimum particle size (pixels) in $x$ direction \\[.3em]
 		
[…]
