+++ user_manual/user_manual.tex
@@ -61,7 +61,7 @@
 
 \begin{minipage}{10cm}
 	{\small 
-	Vesrion: 0.4.0 \\
+	Vesrion: 0.4.1 \\
 	Last updated: \today \\
 	Github repository: \url{https://github.com/ronshnapp/MyPTV} \\
 	Contac: ronshnapp@gmail.com}
