### 2022-05-22 452297d
trying a particle position optimization scheme
