### 2022-05-22 bb6d451
added a function to the workflow for plotting particles images on top of blobs
