### 2022-05-27 ronshnapp/MyPTV#16 @leahmendelson
DOC - typos
I'm going to consolidate a few typos in the user manual into this issue, focusing on those that could confuse a user:

- [x] Section 2.1 of the user manual (Preparation of an experiment folder) step 3 refers to `workfolw.py` instead of `workflow.py` I don't think a user would misunderstand, but given it's the core script for running the code it's worth fixing
- [x] Reference 1 of the user manual has an incomplete journal title
