### 2022-06-14 7dfaba8
not segmentation has an image_start parameter
