+++ example/params_file.yml
@@ -31,6 +31,7 @@
 
 - matching:
     blob_files: blobs_cam1, blobs_cam2, blobs_cam3
+    frame_start: 0
     N_frames: None
     camera_names: cam1, cam2, cam3
     cam_resolution: 1280, 1024
