+++ Readme.md
@@ -1,5 +1,5 @@
-May 16, 2022
-Version: 0.4.1
+June 14, 2022
+Version: 0.4.2
 
 
 # MyPTV
