### 2022-06-25 ronshnapp/MyPTV#16 @ronshnapp
@leahmendelson Thanks, I fixed these typos. I'll leave this issue open in case any other typos are found in the future.
