### 2022-06-25 1eb8b6a
fixed user manual typos
