### 2022-06-25 0bcb321
added camera initial guess reference guide
