### 2022-06-26 88d5f9c
Fixed errors in the user manual calibration section
