### 2022-06-26 d8ff0bd
fixed conda installation instructions in the Readme
