+++ Readme.md
@@ -1,5 +1,5 @@
-June 25, 2022
-Version: 0.4.2
+June 26, 2022
+Version: 0.4.3
 
 
 # MyPTV
