+++ user_manual/user_manual.tex
@@ -59,12 +59,13 @@
 
 \vfill
 
-\begin{minipage}{10cm}
+\begin{minipage}{14cm}
 	{\small 
-	Vesrion: 0.4.2 \\
+	Vesrion: 0.4.3 \\
 	Last updated: \today \\
 	Github repository: \url{https://github.com/ronshnapp/MyPTV} \\
-	Contac: ronshnapp@gmail.com}
+	Get help \& interact with our community: \url{https://github.com/ronshnapp/MyPTV/discussions}\\
+	Contact the developers: ronshnapp@gmail.com}
 \end{minipage}
 
 
