### 2022-06-26 6641c01
JOSS paper corrections based on the reviewers comments
