### 2022-06-26 53d24cb
JOSS paper corrections based on the reviewers comments
