+++ JOSS_paper/paper.md
@@ -27,28 +27,30 @@ Three dimensional particle tracking velocimetry (3D-PTV) is a method that is wid
 
 The application of 3D-PTV relies heavily on computation during the several steps of post-processing the experimental results (see \autoref{fig:setup}). In particular, in many applications researchers study objects that appear in high density in the recorded images, and the images are recorded over extended periods of time at high rates, yielding high volumes of raw data [@Shnapp2019]. Thus, 3D-PTV experiments are inevitably post-processed using specialized computer codes that often need to be tailored to the specific characteristics of the system under investigation. Indeed, the effort the programming of a functioning 3D-PTV software requires might deter inexperienced researchers from employing 3D-PTV, and hinder further development of the method. 
 
-_MyPTV_ is a Python package designed to make 3D-PTV accessible throughout the scientific community, building on the foundation of a previous project. The first open source 3D-PTV software is the _OpenPTV_ project, that was initiated at the early 2000's [@openptv]. The developers of _OpenPTV_ relied on coding in the C language in order to leverage it's high speed of computation for implementing the complex algorithms involved. However, the C language is not accessible to many of the scientists working in the field, so debugging and installation on modern computers can often be challenging, and algorithms from recent years meant to advance the method (e.g. @Ouellette2006; @Xu2008; @Schroeder2015; @Bourgoin2020; @Brizzolara2021) have not yet been implemented in _OpenPTV_. Furthermore, modern computers enable using 3D-PTV with higher level programming tools, while maintaining computational times at a reasonable level. MyPTV_ solves these issues and extends _OpenPTV_ through three principles. First, _MyPTV_ is written exclusively in Python, which is accessible to a wider range of practitioners and widely used in scientific research. This feature allows rapid prototyping and development of the 3D-PTV method, which is crucial for its further development. Second, the dependency on external packages is kept to the bare minimum and includes only an essential set of widely used and properly maintained packages (currently _Numpy_, _Scipy_, *Matplotlib*, _Pandas_, and *Pyyaml*), thus facilitating the maintenance and cross-platform usability without the need for complex deployment phases.  Third, _MyPTV_ extends _OpenPTV_ by including new algorithms for camera calibration, particle tracking, particle segmentation, and trajectory smoothing, that were never implemented in _OpenPTV_. In particular, *MyPTV* includes a novel algorithm for the crucial stereo-matching step that uses time information to prioritize the correspondence of 3D-trackable trajectories. Indeed, now that the code is more accessible, we envision that in the future _MyPTV_  will be further extended by its users to include more developments as they come.   
[…]
