+++ JOSS_paper/bib_myPTV.bib
@@ -1,6 +1,6 @@
 @Article{Virant1997,
-  author    = {M. Virant and T. Dracos},
-  journal   = {Measurement},
+  author    = {Virant, M. and Dracos, T.},
+  journal   = {Measurement Science and Technology},
   title     = {{3D PTV} and its application on {Lagrangian} motion},
   year      = {1997},
   pages     = {1552--1593},
@@ -255,6 +255,72 @@
   owner      = {ron},
   timestamp  = {2018.02.28},
   url        = {http://www.openptv.net/},
+}
+
+@Article{Tan2020,
+  author    = {Tan, Shiyong and Salibindla, Ashwanth and Masuk, Ashik Ullah Mohammad and Ni, Rui},
+  journal   = {Experiments in Fluids},
+  title     = {Introducing {OpenLPT}: new method of removing ghost particles and high-concentration particle shadow tracking},
+  year      = {2020},
+  number    = {2},
+  pages     = {1--16},
+  volume    = {61},
+  doi       = {10.1007/s00348-019-2875-2},
+  publisher = {Springer},
+}
+
+@Article{Schanz2016,
+  author    = {Schanz, Daniel and Gesemann, Sebastian and Schr{\"o}der, Andreas},
+  journal   = {Experiments in fluids},
+  title     = {{Shake-The-Box: Lagrangian} particle tracking at high particle image densities},
+  year      = {2016},
+  number    = {5},
+  pages     = {1--27},
+  volume    = {57},
+  doi       = {10.1007/s00348-016-2157-1},
+  publisher = {Springer},
+}
+
+@Article{Wieneke2013,
+  author    = {Wieneke, B.},
+  journal   = {Measurement Science and Technology},
+  title     = {Iterative reconstruction of volumetric particle distribution},
+  year      = {2013},
+  number    = {2},
+  pages     = {024008},
+  volume    = {24},
+  doi       = {10.1088/0957-0233/24/2/024008},
+  publisher = {IOP Publishing},
+}
+
+@Article{Sbalzarini2005,
+  author    = {Sbalzarini, I. F. and Koumoutsakos, P.},
+  journal   = {Journal of structural biology},
+  title     = {Feature point tracking and trajectory analysis for video imaging in cell biology},
+  year      = {2005},
+  number    = {2},
+  pages     = {182--195},
+  volume    = {151},
+  doi       = {10.1016/j.jsb.2005.06.002},
+  publisher = {Elsevier},
+}
+
+@Article{Heyman2019,
+  author    = {Heyman, J.},
+  journal   = {Computers \& Geosciences},
+  title     = {{TracTrac: A} fast multi-object tracking algorithm for motion estimation},
+  year      = {2019},
+  pages     = {11--18},
+  volume    = {128},
+  doi       = {10.1016/j.cageo.2019.03.007},
+  publisher = {Elsevier},
+}
+
+@Article{Allan2021,
+  author = {Allan, D. B. and Caswell, T. and Keim, N. C. and van der Wel, C. M. and Verweij, R. W.},
+  title  = {soft-matter/trackpy: Trackpy v0.5.0},
+  year   = {2021},
+  doi    = {10.5281/zenodo.4682814},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
