+++ JOSS_paper/paper.md
@@ -27,7 +27,7 @@ Three dimensional particle tracking velocimetry (3D-PTV) is a method that is wid
 
 The application of 3D-PTV relies heavily on computation during the several steps of post-processing the experimental results (see \autoref{fig:setup}). In particular, in many applications researchers study objects that appear in high density in the recorded images, and the images are recorded over extended periods of time at high rates, yielding high volumes of raw data [@Shnapp2019]. Thus, 3D-PTV experiments are inevitably post-processed using specialized computer codes that often need to be tailored to the specific characteristics of the system under investigation. Indeed, the effort the programming of a functioning 3D-PTV software requires might deter inexperienced researchers from employing 3D-PTV, and hinder further development of the method. 
 
-_MyPTV_ is a Python package designed to make 3D-PTV accessible throughout the scientific community, building on the foundation of a previous project. The first open source 3D-PTV software is the _OpenPTV_ project, that was initiated in the early 2000's [@openptv] (although the algorithms and the code goes more than a decade earlier). The developers of _OpenPTV_ relied on coding in the C language in order to leverage it's high speed of computation for implementing the complex algorithms involved. Another recent open source project, the C++ written OpenLPT [@Tan2020], enables users to employ the shake-the-box algorithm [@Schanz2016] in 3D particle tracking experiments. While these two important projects enable using high-performance 3D-PTV, the C and C++ languages in which they are written are not accessible to many of the scientists working in the field. Therefore, debugging and installation can often be challenging as computers and operating systems evolve with time. Furthermore, new algorithms such as the ones introduced in [@Ouellette2006; @Xu2008; @Wieneke2013; @Schroeder2015; @Schanz2016; @Bourgoin2020; @Tan2020] have not yet been implemented in _OpenPTV_, and the development of novel algorithms, e.g. [@Brizzolara2021], is more restrictive in these low level coding languages. And yet, modern computers enable using 3D-PTV with higher level programming tools while maintaining computational times at a reasonable level. In addition to that, several tools exist for particle tracking in two dimensions (e.g. [@Sbalzarini2005; @Heyman2019; @Allan2021]), however they do not include 3D models and thus are limited to describing the motion in two dimensions. 
+_MyPTV_ is a Python package designed to make 3D-PTV accessible throughout the scientific community, building on the foundation of a previous project. The first open source 3D-PTV software is the _OpenPTV_ project, that was initiated in the early 2000's [@openptv] (although the algorithms and the code goes more than a decade earlier). The developers of _OpenPTV_ relied on coding in the C language in order to leverage it's high speed of computation for implementing the complex algorithms involved. Another recent open source project, the C++ written _OpenLPT_ [@Tan2020], enables users to employ the shake-the-box algorithm [@Schanz2016] in 3D particle tracking experiments. While these two important projects enable using high-performance 3D-PTV, the C and C++ languages in which they are written are not accessible to many of the scientists working in the field. Therefore, debugging and installation can often be challenging as computers and operating systems evolve with time. Furthermore, new algorithms such as the ones introduced in [@Ouellette2006; @Xu2008; @Wieneke2013; @Schroeder2015; @Schanz2016; @Bourgoin2020; @Tan2020] have not yet been implemented in _OpenPTV_, and the development of novel algorithms, e.g. [@Brizzolara2021], is more restrictive in these low level coding languages. And yet, modern computers enable using 3D-PTV with higher level programming tools while maintaining computational times at a reasonable level. In addition to that, several tools exist for particle tracking in two dimensions [@Sbalzarini2005; @Heyman2019; @Allan2021], however they do not include 3D models and thus are limited to describing the motion in two dimensions. 
 
[…]
