### 2022-06-26 #4398 @ronshnapp
Hi @jgostick and @leahmendelson.
I went through the detailed comments and issues raised in the review and addressed them all (@leahmendelson, thank you very much, I truly appreciate your effort and value your comments). The revision to the code, manual, and readme can be found in the master branch. The revision to the paper can be found in the JOSS_paper branch. The changes I made are listed as comments in the issues raised.
Kindly let me know what I should do in order to proceed.
