### 2022-06-26 53d5c60
improved runtime of the stitching algorithm
