### 2022-06-27 c09b9a3
added gif to the readme
