### 2022-06-27 f06be3f
added logo to manual
