### 2022-07-09 #4398 @ronshnapp
Hi @jgostick, I wanted to ask if there is any status update on the paper?
