+++ myptv/cal_gui.py
@@ -1,0 +1,325 @@
+# -*- coding: utf-8 -*-
+"""
+Created on April 23, 2022
+
+@author: Ron
+
+A gui for manually marking points on a calibration target. 
+"""
+
+
+#from PIL import Image, ImageTk
+from tkinter import Label, LabelFrame, Entry, Tk, Button
+from matplotlib.pyplot import subplots, show, imread
+
+
+class cal_gui(object):
+    '''
+    This is a Tkinter based graphical user interface that can be used to mark
+    points on a static image, give their lab space coordinates, and then save
+    the data as a text file in the format used by the calibration processes.
+    '''
+    
+    
+    def __init__(self, calibrate_obj=None, cal_image=None):
+        '''
+        input: 
+        
+        calibrate_obj - An instance of the calibrate class, ready to perform
+                        calibration.
+        
+        '''
+        self.calibrate_obj = calibrate_obj
+        self.cal_image = cal_image
+        self.err = self.calibrate_obj.mean_squared_err()
+        
+        # set the window
+        self.root = Tk()
+        self.root.geometry('400x380')
+        self.root.resizable(0,0)
+        self.root.title('MyPTV: calibration GUI')
+
+
+        # =============================
+        # BUTTONS:
+        
+        # buttons frame
+        button_label = LabelFrame(self.root, padx=10, pady=3, width=100)
+        button_label.grid(row=0, column=0, padx=(10), pady=3, sticky='nsew')
+        
+
+        # Buttons frame
+        button_frame1 = Label(button_label)
+        button_frame1.grid(row=0, column=0, columnspan=1, sticky='nsw', padx=2, 
+                          pady=3)
+        
+        
+        # external params calibration button
+        external_button = Button(button_frame1, text='External calibration', 
+                                command = self.external_calibration, 
+                                padx=2, pady=4, width=20)
+        external_button.grid(row=0, column=0, padx=10, pady=2, sticky='ew')
+        
+        # non-linear correction term
+        fine_cal_button = Button(button_frame1, text='Fine calibration', 
+                                command = self.fine_calibration, 
+                                padx=2, pady=4, width=20)
+        fine_cal_button.grid(row=1, column=0, padx=10, pady=2, sticky='ew')
+        
+        # plot calibration
+        plot_button = Button(button_frame1, text='Plot calibration', 
+                                command = self.plot_calibration, 
+                                padx=2, pady=4, width=20)
+        plot_button.grid(row=2, column=0, padx=10, pady=2, sticky='ew')
+        
+        
+        # save points button
+        save_button = Button(button_frame1, text='Save', 
+                                command = self.Save, padx=10, pady=4, width=15) 
+        save_button.grid(row=0, column=1, padx=10, pady=2, sticky='ew')
+        
+        # quit button
+        quit_button = Button(button_frame1, text='Quit', 
+                                command = self.Quit, padx=10, pady=4) 
+        quit_button.grid(row=1, column=1, padx=10, pady=2, sticky='ew')
+        
+        
+        
+        
+        
+        
+        
+        # =============================
+        # CALIBRATOR STATUS:
+        
+            
+        status_label = LabelFrame(self.root, padx=10, pady=4, width=100)
+        status_label.grid(row=1, column=0, padx=(10), pady=4, sticky='nsew')
+        
[…]
