### 2022-07-13 a2e50da
extended calibration gui with manual parameter changing
