+++ myptv/cal_gui.py
@@ -9,7 +9,7 @@ A gui for manually marking points on a calibration target.
 
 
 #from PIL import Image, ImageTk
-from tkinter import Label, LabelFrame, Entry, Tk, Button
+from tkinter import Label, LabelFrame, Entry, Tk, Button, Checkbutton, IntVar
 from matplotlib.pyplot import subplots, show, imread
 
 
@@ -31,11 +31,10 @@ class cal_gui(object):
         '''
         self.calibrate_obj = calibrate_obj
         self.cal_image = cal_image
-        self.err = self.calibrate_obj.mean_squared_err()
         
         # set the window
         self.root = Tk()
-        self.root.geometry('400x380')
+        self.root.geometry('400x480')
         self.root.resizable(0,0)
         self.root.title('MyPTV: calibration GUI')
 
@@ -66,11 +65,20 @@ class cal_gui(object):
                                 padx=2, pady=4, width=20)
         fine_cal_button.grid(row=1, column=0, padx=10, pady=2, sticky='ew')
         
+        
+        # grab manual button
+        grab_button = Button(button_frame1, text='Grab manual', 
+                                command = self.grab_manual, 
+                                padx=2, pady=4, width=14)
+        grab_button.grid(row=2, column=0, padx=10, pady=2, sticky='ew')
+        
+        
+        
         # plot calibration
         plot_button = Button(button_frame1, text='Plot calibration', 
                                 command = self.plot_calibration, 
                                 padx=2, pady=4, width=20)
-        plot_button.grid(row=2, column=0, padx=10, pady=2, sticky='ew')
+        plot_button.grid(row=3, column=0, padx=10, pady=2, sticky='ew')
         
         
         # save points button
@@ -99,9 +107,31 @@ class cal_gui(object):
         self.status = Label(status_label, text='Status:', padx=2, pady=2)
         self.status.grid(row=0, column=0, rowspan=1, sticky='nw', padx=2, pady=2)
         self.status_show = Label(status_label, text='wating for action', padx=2, pady=2,
-                                 width=50, anchor='w', fg='green')
+                                 width=40, anchor='w', fg='green')
         self.status_show.grid(row=0, column=1, rowspan=1, sticky='nw', padx=2, pady=2)
         
+        
+        
+        self.camera = Label(status_label, text='Camera:', padx=2, pady=2)
+        self.camera.grid(row=1, column=0, rowspan=1, sticky='nw', padx=2, pady=2)
+        
+        cam = 'camera name'
+        if self.calibrate_obj is not None:
+            cam = self.calibrate_obj.camera.name
+        
+        self.camera_show = Label(status_label, text=cam, padx=2, pady=2,
+                                 width=40, anchor='w')
+        self.camera_show.grid(row=1, column=1, rowspan=1, sticky='nw', padx=2, pady=2)
+
+
+
+        self.manual_var = IntVar()
+        self.manual_state = Checkbutton(status_label, padx=2, pady=2,
+                                        command=self.manual_onof, 
+                                        variable = self.manual_var)
+        self.manual_state.grid(row=2, column=1, rowspan=1, sticky='nw', padx=2, pady=2)
+        self.manual = Label(status_label, text='Manual:', padx=2, pady=2)
+        self.manual.grid(row=2, column=0, rowspan=1, sticky='nw', padx=2, pady=2)
         
         
         
@@ -126,19 +156,21 @@ class cal_gui(object):
         self.x_input = Entry(pos_dashboard, width=14)
         self.x_input.insert(0,'0.0')
[…]
