### 2022-07-16 #4398 @jgostick
@ronshnapp I think this package is good to go and I will begin the process of accepting it.  Well done! There are a few steps left, as outlined in the checklist below:

Here is a list of what to do next:

- [ ] Set article DOI with ``@editorialbot set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@editorialbot recommend-accept``
