### 2022-07-16 #4398 @ronshnapp
Hi @jgostick,
Thank you very much, I'm very excited to see MyPTV advancing with its first publication!

The version tag is V0.4.3
The DOI on Zenodo in: 10.5281/zenodo.6846969
I confirm checking that the author list and affiliations are correct on the paper and on Zenodo.
