### 2022-07-26 6c4fb16
added save warning for existing files in the segmentation
