### 2022-07-26 95f6cc4
added save files warning to the smoothing workflow
