+++ example/workflow.py
@@ -729,6 +729,7 @@ class workflow(object):
         '''
         from numpy import loadtxt
         from myptv.traj_smoothing_mod import smooth_trajectories
+        from os import getcwd, listdir
         
         # fetching the smoothing parameters
         trajectory_file = self.get_param('smoothing', 'trajectory_file')
@@ -746,8 +747,20 @@ class workflow(object):
         
         # saving the data
         if save_name is not None:
-            print('\n', 'Saving the smoothed data (%s).'%save_name)
-            sm.save_results(save_name)
+            cwd_ls = listdir(getcwd())
+            if save_name in cwd_ls:
+                print('\n The file name "%s" already exists in'%save_name)
+                print(' the working directory. Should I save anyways?')
+                usr = input('(1=yes, else=no)')
+                if usr == '1':
+                    print('\n', 'Saving the smoothed data (%s).'%save_name)
+                    sm.save_results(save_name)
+                else:
+                    print('\n', 'Skipped saving file.')
+            
+            else:
+                print('\n', 'Saving the smoothed data (%s).'%save_name)
+                sm.save_results(save_name)
         
         print('\n', 'Done.')
         
