### 2022-07-26 621731f
added file save warning to stitching
