### 2022-07-26 67e4ce3
several biug fixes in the stitching mod and attempt to make it slightly faster
