+++ myptv/traj_stitching_mod.py
@@ -17,6 +17,7 @@ from numpy import array, gradient, dot, savetxt
 from numpy import append as NPappend
 from myptv.utils import fit_polynomial
 from myptv.traj_smoothing_mod import smooth_traj_poly
+from itertools import groupby
 
 
 class traj_stitching(object):    
@@ -62,7 +63,8 @@ class traj_stitching(object):
         i - int, trajectory number
         '''
         key = lambda s: s[-1]
-        tr = array(sorted(self.traj_list[self.traj_list[:,0] == i], key=key))
+        #tr = array(sorted(self.traj_list[self.traj_list[:,0] == i], key=key))
+        tr = array(sorted(self.traj_list[self.index_id_hash[i],:], key=key))
         return tr
 
 
@@ -131,9 +133,28 @@ class traj_stitching(object):
         
         # make a dictionary whose keys are trajectory ids and values are
         # the trajectories themselves
+# =============================================================================
+#         Legacy code:
+#
+#         traj_dic = {}
+#         for id_ in traj_ids:
+#             tr = self.get_traj(id_)
+#             v = tr[:,4:7]
+#             if (v==0).all():
+#                 x = tr[:,1:4]
+#                 v = gradient(x, axis=0)
+#                 a = gradient(v, axis=0)
+#                 tr[:,4:7] = v
+#                 tr[:,7:10] = a
+#                 
+#             traj_dic[id_] = tr 
+# =============================================================================
+        
         traj_dic = {}
-        for id_ in traj_ids:
-            tr = self.get_traj(id_)
+        trajs = [array(sorted(list(g), key=lambda a: a[-1])) for k,g in 
+                 groupby(self.traj_list, key=lambda x: x[0]) if k!=-1]
+        for tr in trajs:
+            id_ = tr[0,0]
             v = tr[:,4:7]
             if (v==0).all():
                 x = tr[:,1:4]
@@ -141,7 +162,6 @@ class traj_stitching(object):
                 a = gradient(v, axis=0)
                 tr[:,4:7] = v
                 tr[:,7:10] = a
-                
             traj_dic[id_] = tr 
         
         # make a dicionary whose keys are frame numbers and values are lists
@@ -179,18 +199,7 @@ class traj_stitching(object):
                 
             count += 1
         print('') 
-        
-        # make the list of d_ij
-        # traj_ids = sorted(traj_ids, key=lambda i: traj_dic[i][0,-1])
-        # self.dij_list = []
-        # for i in range(len(traj_ids)):
-        #     for j in range(len(traj_ids)):
-        #         id_i, id_j = traj_ids[i], traj_ids[j]
-        #         traj_i, traj_j = traj_dic[id_i], traj_dic[id_j]
-        #         dij = self.calc_d(traj_i, traj_j)
-        #         if dij != -1:
-        #             dt = traj_j[0,-1] - traj_i[-1,-1]
-        #             self.dij_list.append((id_i, id_j, dt, dij))
+
         
     
     
@@ -232,7 +241,11 @@ class traj_stitching(object):
         '''
         connected_traj_i = []
         connected_traj_j = []
+        
+        count = 1
         for i,j,dt,dij in self.dij_list:
+            print('', end='\r')
+            print(' %d / %d'%(count, len(self.dij_list)), end='\r')
             
             tr_i = self.get_traj(i)
             tr_j = self.get_traj(j)
@@ -253,41 +266,59 @@ class traj_stitching(object):
             # the time needed to interpolate
             tm_interp = [float(i) for i in range(1,int(dt))]
             
-            # in each direction fit a polynomial and fill missing samples
[…]
