### 2022-07-30 #4398 @jgostick
@ronshnapp very sorry for the delays on this review.  Please don't let it reflect badly on JOSS, we jsut had a reviewer drop out, and I have had full teaching load this summer, plus several conferences, etc, so it's been hectic.  Anyway, congrats, this looks like a great package.  An editor-in-chief will be by soon to make it official.
