### 2022-07-30 #4398 @ronshnapp
@jgostick There's nothing to be sorry about. I think JOSS is a great project and the fact people such a you do so much to keep it going is incredible. I will definitely consider chipping in as a reviewer despite a slight impostor syndrome. Thank you very much for managing this process!
