### 2022-07-30 #4398 @Kevin-Mattheus-Moerman
@ronshnapp

I am an AEiC for JOSS and here to help process the final steps for acceptance. I have studied this review issues, your repository, and the paper, and have some issues below that require your attention.

- [ ] Please manually edit the ZENODO archive such that the title matches that of the paper.
- [ ] The ZENODO archive currently lists "Other (open)" as the license. I believe you can manually edit it to refer to your MIT license.

I have read your paper and have the following comments:
- [ ] Please add the city and country to your affiliation.
- [ ] Can you check if `Applications of 3D-PTV abound...`, should read `Applications of 3D-PTV are abound...`.
- [ ] Check `the effort the programming` which should probably be `the effort of programming`.
- [ ] When you say `building on the foundation of a previous project`, can you mention/cite/refer to that project? Sorry if I missed it.
- [ ] Check `leverage it’s high speed`, which should be `leverage its high speed`
- [ ] The paper says `MyPTV, currently in version 0.2, contains...`, yet the review here is for version V0.4.3. Should the text in the paper be amended?
- [ ] It looks like it would be better if `MyPTV had been tested in a series..` is written `MyPTV has been tested in a series...`.
- [ ] It looks like `MyPTV includes several helpful tools to ensure the software user friendly...` should be rewritten e.g. as `MyPTV includes several helpful tools to ensure the software's user friendliness...` or use something like `to ensure the user friendliness of the software`.

Please work on the above and we can proceed with the next steps. Thanks.
