### 2022-07-30 d8a28fe
implemented corrections to the JOSS paper
