### 2022-07-30 22426bf
added JOSS published paper and updated readme
